Make PamHandle2 work!
I've had to look into libpam to see that the "struct pam_conv *" passed as an
argument, to pam_start as well as pam_set_item, is actually duplicated.

Also, you cannot pam_set_item() a PAM_AUTHTOK if you're not a module! The old
code therefore got this wrong. Instead, here, we overwrite the pam_conv before
authenticating so that the password be included with it and use it during our
"interview".